fixed user buy book

+++ backend/src/book/book.service.ts
@@ -1,14 +1,15 @@
 import { HttpException, Injectable } from '@nestjs/common';
 import { InjectModel } from "@nestjs/mongoose";
-import { Book } from "./book.schema";
-import { Model } from "mongoose";
+import { Book, BookDocument } from "./book.schema";
+import mongoose, { Model } from "mongoose";
 import { CreateBookDto } from "./dto/create-book.dto";
 import { UpdateBookDto } from "./dto/update-book.dto";
 import {v4} from "uuid"
+import { BuyBookDto } from "./dto/buy-book.dto";
 
 @Injectable()
 export class BookService {
-  constructor(@InjectModel(Book.name) private bookModel: Model<Book>) {}
+  constructor(@InjectModel(Book.name) private bookModel: Model<BookDocument>) {}
 
   async create(createBookDto: CreateBookDto) {
     if (!createBookDto.title) {
@@ -21,20 +22,20 @@ export class BookService {
       throw new HttpException('missing image link', 422);
     }
     const book = new this.bookModel(createBookDto);
-    book._id = v4()
+    book._id = new mongoose.Types.ObjectId()
     await book.save()
     return book;
   }
 
   async findAll(name: string) {
     if(!name){
-      return await this.bookModel.find();
+      return await this.bookModel.find().populate("book");
     }
-    return await this.bookModel.find({title: name})
+    return await this.bookModel.find({title: name}).populate("book")
   }
 
   async findOne(id: string) {
-    return await this.bookModel.findById(id);
+    return await this.bookModel.findById(id).populate("book").exec();
   }
 
   async update(id: string, updateBookDto: UpdateBookDto): Promise<Book> {
@@ -49,6 +50,20 @@ export class BookService {
   }
 
   async delete(id: string) {
+    if(!id){
+      throw new HttpException('id is required', 400);
+    }
     return this.bookModel.findByIdAndDelete(id);
+  }
+
+  async findToBuy(books: BuyBookDto[]): Promise<Book[]>{
+     let  booksList = []
+      books.forEach(async (book: BuyBookDto) => {
+      let foundBook = await this.bookModel.findById(book._id)
+      if(foundBook){
+        booksList.push(foundBook)
+      }
+    })
+    return booksList
   }
 }
